chore(api): move generate-username usecase  into prescription context
