+++ api/src/prescription/organization-learner/domain/usecases/generate-username.js
@@ -1,10 +1,10 @@
 import lodash from 'lodash';
 
-import { STUDENT_RECONCILIATION_ERRORS } from '../../../src/shared/domain/constants.js';
+import { STUDENT_RECONCILIATION_ERRORS } from '../../../../shared/domain/constants.js';
 import {
   OrganizationLearnerAlreadyLinkedToUserError,
   OrganizationLearnerNotFound,
-} from '../../../src/shared/domain/errors.js';
+} from '../../../../shared/domain/errors.js';
 
 const { find, get } = lodash;
